# api/app/docs/swagger.yaml
definitions:
...
  dashboard_internal_feature_payroll_summary.LatestRunDTO:
    properties:
      employeeCount:
        type: integer
      id:
        type: string
      payrollMonthDate:
        type: string
      status:
        type: string
      totalNetPay:
        type: number
      totalPf:
        type: number
      totalSso:
        type: number
      totalTax:
        type: number
    type: object
  dashboard_internal_feature_payroll_summary.MonthlyBreakdownDTO:
    properties:
      month:
        type: string
      netPay:
        type: number
      pf:
        type: number
      sso:
        type: number
      tax:
        type: number
    type: object
  dashboard_internal_feature_payroll_summary.PayrollSummaryResponse:
    properties:
      latestRun:
        $ref: '#/definitions/dashboard_internal_feature_payroll_summary.LatestRunDTO'
      monthlyBreakdown:
        items:
          $ref: '#/definitions/dashboard_internal_feature_payroll_summary.MonthlyBreakdownDTO'
        type: array
      year:
        type: integer
      yearlyTotals:
        $ref: '#/definitions/dashboard_internal_feature_payroll_summary.YearlyTotalsDTO'
    type: object
  dashboard_internal_feature_payroll_summary.YearlyTotalsDTO:
    properties:
      totalNetPay:
        type: number
      totalPf:
        type: number
      totalSso:
        type: number
      totalTax:
        type: number
    type: object
...
  hrms_modules_dashboard_internal_feature_attendance_summary.AttendanceBreakdown:
    properties:
      lateCount:
        type: integer
      lateMinutes:
        type: number
      leaveDayCount:
        type: integer
      leaveDays:
        type: number
      leaveDoubleCount:
        type: integer
      leaveDoubleDays:
        type: number
      leaveHours:
        type: number
      leaveHoursCount:
        type: integer
      otCount:
        type: integer
      otHours:
        type: number
      period:
        type: string
    type: object
  hrms_modules_dashboard_internal_feature_attendance_summary.AttendanceSummaryResponse:
    properties:
      breakdown:
        items:
          $ref: '#/definitions/hrms_modules_dashboard_internal_feature_attendance_summary.AttendanceBreakdown'
        type: array
      period:
        properties:
          endDate:
            type: string
          startDate:
            type: string
        type: object
      totals:
        $ref: '#/definitions/hrms_modules_dashboard_internal_feature_attendance_summary.AttendanceTotals'
    type: object
  hrms_modules_dashboard_internal_feature_attendance_summary.AttendanceTotals:
    properties:
      lateCount:
        type: integer
      lateMinutes:
        type: number
      leaveDayCount:
        type: integer
      leaveDays:
        type: number
      leaveDoubleCount:
        type: integer
      leaveDoubleDays:
        type: number
      leaveHours:
        type: number
      leaveHoursCount:
        type: integer
      otCount:
        type: integer
      otHours:
        type: number
    type: object
...
  hrms_modules_dashboard_internal_feature_employee_summary.DepartmentCountDTO:
    properties:
      count:
        type: integer
      departmentId:
        type: string
      departmentName:
        type: string
    type: object
  hrms_modules_dashboard_internal_feature_employee_summary.EmployeeSummaryResponse:
    properties:
      activeEmployees:
        type: integer
      byDepartment:
        items:
          $ref: '#/definitions/hrms_modules_dashboard_internal_feature_employee_summary.DepartmentCountDTO'
        type: array
      fullTimeCount:
        type: integer
      newThisMonth:
        type: integer
      partTimeCount:
        type: integer
      terminatedThisMonth:
        type: integer
      totalEmployees:
        type: integer
    type: object
...
  hrms_modules_payoutpt_internal_feature_create.Command:
    properties:
      eb: {}
      employeeId:
        type: string
      repo:
        $ref: '#/definitions/hrms_modules_payoutpt_internal_repository.Repository'
      tx: {}
      worklogIds:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - employeeId
    - worklogIds
    type: object
  hrms_modules_payoutpt_internal_feature_create.Response:
    properties:
      payout:
        $ref: '#/definitions/hrms_modules_payoutpt_internal_repository.Payout'
    type: object
...
  hrms_modules_payoutpt_internal_repository.Payout:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      employeeId:
        type: string
      hourlyRate:
        type: number
      id:
        type: string
      itemCount:
        type: integer
      paidAt:
        type: string
      paidBy:
        type: string
      status:
        type: string
      totalHours:
        type: number
      updatedAt:
        type: string
    type: object
  hrms_modules_payoutpt_internal_repository.PayoutItem:
    properties:
      id:
        type: string
      totalHours:
        type: number
      workDate:
        type: string
      worklogId:
        type: string
    type: object
  hrms_modules_payoutpt_internal_repository.Repository:
    type: object
...
  hrms_modules_payrollrun_internal_dto.RunTotals:
    properties:
      totalDeduction:
        type: number
      totalIncome:
        type: number
      totalNetPay:
        type: number
      totalProvidentFund:
        type: number
      totalSocialSecurity:
        type: number
      totalTax:
        type: number
    type: object
...
  payoutpt_internal_feature_get.Response:
    properties:
      items:
        items:
          $ref: '#/definitions/hrms_modules_payoutpt_internal_repository.PayoutItem'
        type: array
      payout:
        $ref: '#/definitions/hrms_modules_payoutpt_internal_repository.Payout'
    type: object
  payoutpt_internal_feature_list.Meta:
    properties:
      currentPage:
        type: integer
      totalItems:
        type: integer
      totalPages:
        type: integer
    type: object
  payoutpt_internal_feature_list.Response:
    properties:
      data:
        items:
          $ref: '#/definitions/hrms_modules_payoutpt_internal_repository.Payout'
        type: array
      meta:
        $ref: '#/definitions/payoutpt_internal_feature_list.Meta'
    type: object
  payoutpt_internal_feature_pay.Payout:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      employeeId:
        type: string
      hourlyRate:
        type: number
      id:
        type: string
      itemCount:
        type: integer
      paidAt:
        type: string
      paidBy:
        type: string
      status:
        type: string
      totalHours:
        type: number
      updatedAt:
        type: string
    type: object
...
  payrollrun_internal_feature_get.Response:
    properties:
      approvedAt:
        type: string
      approvedBy:
        type: string
      bonusYear:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      orgProfileSnapshot:
        additionalProperties: true
        type: object
      payDate:
        type: string
      payrollMonthDate:
        type: string
      periodStartDate:
        type: string
      socialSecurityRateEmployee:
        type: number
      socialSecurityRateEmployer:
        type: number
      status:
        type: string
      totalEmployees:
        type: integer
      totalNetPay:
        type: number
      totals:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.RunTotals'
      updatedAt:
        type: string
    type: object
...
  payrollrun_internal_feature_update.Command:
    properties:
      payDate:
        type: string
      status:
        enum:
        - pending
        - approved
        type: string
    type: object
  payrollrun_internal_feature_update.Response:
    properties:
      approvedAt:
        type: string
      approvedBy:
        type: string
      bonusYear:
        type: integer
      createdAt:
        type: string
      id:
        type: string
      message:
        type: string
      orgProfileSnapshot:
        additionalProperties: true
        type: object
      payDate:
        type: string
      payrollMonthDate:
        type: string
      periodStartDate:
        type: string
      socialSecurityRateEmployee:
        type: number
      socialSecurityRateEmployer:
        type: number
      status:
        type: string
      totalEmployees:
        type: integer
      totalNetPay:
        type: number
      totals:
        $ref: '#/definitions/hrms_modules_payrollrun_internal_dto.RunTotals'
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: HR Payroll REST API
  title: HR Payroll API
  version: "1.0"
paths:
...
  /dashboard/attendance-summary:
    get:
      description: Get attendance statistics with optional grouping and filtering
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        required: true
        type: string
      - description: 'Group by: month or day (default: month)'
        in: query
        name: groupBy
        type: string
      - description: Filter by department ID
        in: query
        name: departmentId
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_dashboard_internal_feature_attendance_summary.AttendanceSummaryResponse'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get attendance summary
      tags:
      - Dashboard
...
  /dashboard/employee-summary:
    get:
      description: Get aggregated employee statistics
      parameters:
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hrms_modules_dashboard_internal_feature_employee_summary.EmployeeSummaryResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get employee summary
      tags:
      - Dashboard
...
  /dashboard/payroll-summary:
    get:
      description: Get payroll statistics including latest run and yearly totals
      parameters:
      - description: 'Year (default: current year)'
        in: query
        name: year
        type: integer
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dashboard_internal_feature_payroll_summary.PayrollSummaryResponse'
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get payroll summary
      tags:
      - Dashboard
...
  /payouts/pt:
    get:
      parameters:
      - description: filter by employee
        in: query
        name: employeeId
        type: string
      - description: to_pay|paid|all
        in: query
        name: status
        type: string
      - description: filter by start date (YYYY-MM-DD)
        in: query
        name: startDate
        type: string
      - description: filter by end date (YYYY-MM-DD)
        in: query
        name: endDate
        type: string
      - description: page
        in: query
        name: page
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payoutpt_internal_feature_list.Response'
      security:
      - BearerAuth: []
      summary: List part-time payouts
      tags:
      - Part-Time Payout
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hrms_modules_payoutpt_internal_feature_create.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/hrms_modules_payoutpt_internal_feature_create.Response'
      security:
      - BearerAuth: []
      summary: Create part-time payout
      tags:
      - Part-Time Payout
  /payouts/pt/{id}:
    delete:
      description: ยกเลิกใบเบิกจ่ายที่ยังไม่ได้จ่าย (status=to_pay)
      parameters:
      - description: payout id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Cancel part-time payout
      tags:
      - Part-Time Payout
    get:
      parameters:
      - description: payout id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payoutpt_internal_feature_get.Response'
      security:
      - BearerAuth: []
      summary: Get payout detail
      tags:
      - Part-Time Payout
  /payouts/pt/{id}/pay:
    post:
      parameters:
      - description: payout id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payoutpt_internal_feature_pay.Payout'
      security:
      - BearerAuth: []
      summary: Mark payout paid
      tags:
      - Part-Time Payout
...
  /payroll-runs/{id}:
    delete:
      description: ลบงวดเงินเดือน (เฉพาะ pending)
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Delete payroll run
      tags:
      - Payroll Run
    get:
      description: ดูข้อมูลงวดเงินเดือน
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payrollrun_internal_feature_get.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Get payroll run detail
      tags:
      - Payroll Run
    patch:
      consumes:
      - application/json
      description: แก้ payDate หรืออนุมัติ run
      parameters:
      - description: run id
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payrollrun_internal_feature_update.Command'
      - description: Company ID
        in: header
        name: X-Company-ID
        type: string
      - description: Branch ID
        in: header
        name: X-Branch-ID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payrollrun_internal_feature_update.Response'
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
      security:
      - BearerAuth: []
      summary: Update/approve payroll run
      tags:
      - Payroll Run
...
securityDefinitions:
  BearerAuth:
    description: Add "Bearer <access token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"